fix(kafka): Fix data type for kafka maxPollIntervalMs
Refs: FST-85

+++ folio-service-tools-spring-dev/src/main/java/org/folio/spring/tools/kafka/FolioKafkaProperties.java
@@ -70,7 +70,7 @@ public class FolioKafkaProperties {
     /**
      * Max processing time for a single poll.
      */
-    private Long maxPollIntervalMs;
+    private Integer maxPollIntervalMs;
   }
 
   @Data
